feat(post): add rudimentary CRUD for posts

+++ src/post/post.controller.ts
@@ -6,18 +6,26 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  Request,
+  NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { PostService } from './post.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { ApiBearerAuth } from '@nestjs/swagger';
 
 @Controller('post')
 export class PostController {
   constructor(private readonly postService: PostService) {}
 
   @Post()
-  create(@Body() createPostDto: CreatePostDto) {
-    return this.postService.create(createPostDto);
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  create(@Body() createPostDto: CreatePostDto, @Request() req) {
+    return this.postService.create(createPostDto, req.user.id);
   }
 
   @Get()
@@ -31,12 +39,28 @@ export class PostController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updatePostDto: UpdatePostDto) {
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async update(
+    @Param('id') id: string,
+    @Body() updatePostDto: UpdatePostDto,
+    @Request() req,
+  ) {
+    const post = await this.postService.findOne(+id);
+    if (!post) throw new NotFoundException();
+    const userId = req.user.id;
+    if (userId != post.organizer.id) throw new ForbiddenException();
     return this.postService.update(+id, updatePostDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  async remove(@Param('id') id: string, @Request() req) {
+    const post = await this.postService.findOne(+id);
+    if (!post) throw new NotFoundException();
+    const userId = req.user.id;
+    if (userId != post.organizer.id) throw new ForbiddenException();
     return this.postService.remove(+id);
   }
 }
